[EE] Move MultiSeed behaviour into dedicated packages
