+++ api/cmd/kubermatic-operator-util/main_ce.go
@@ -1,0 +1,11 @@
+// +build !ee
+
+package main
+
+import (
+	"github.com/urfave/cli"
+)
+
+func extraCommands() []cli.Command {
+	return nil
+}
