+++ api/pkg/ee/conversion/helm.go
@@ -1,3 +1,5 @@
+// +build ee
+
 package conversion
 
 import (
@@ -10,8 +12,8 @@ import (
 	kubermaticv1 "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
 	"github.com/kubermatic/kubermatic/api/pkg/crd/migrations/util"
 	operatorv1alpha1 "github.com/kubermatic/kubermatic/api/pkg/crd/operator/v1alpha1"
+	"github.com/kubermatic/kubermatic/api/pkg/ee/provider"
 	"github.com/kubermatic/kubermatic/api/pkg/features"
-	"github.com/kubermatic/kubermatic/api/pkg/provider"
 	"github.com/kubermatic/kubermatic/api/pkg/provider/kubernetes"
 	"github.com/kubermatic/kubermatic/api/pkg/resources"
