+++ api/pkg/ee/conversion/helm_test.go
@@ -1,3 +1,5 @@
+// +build ee
+
 package conversion
 
 import (
