+++ api/pkg/ee/provider/conversions_test.go
@@ -1,3 +1,5 @@
+// +build ee
+
 package provider
 
 import (
